Added an onTouchEvent making the ball's X-axis move when touching the screen. Changed values from int to float in Sprite.

+++ app/src/main/java/jonathan/balljumper/classes/Panel.java
@@ -7,7 +7,7 @@ import android.graphics.Paint;
  */
 
 public class Panel extends Sprite {
-    public Panel(int x, int y, int width, int height, int color) {
+    public Panel(float x, float y, int width, int height, int color) {
         super(x, y, width, height, color);
     }
 }
